Add files via upload
# 🎬 Day 29 – Movie Rating Prediction using Collaborative Filtering

This project predicts how a user might rate a movie they haven’t seen yet using **Collaborative Filtering** techniques. It’s a fundamental project in the world of recommender systems.

---

## 📌 Overview

The goal is to build a **movie recommender model** that predicts user ratings based on the preferences of similar users or items. This is achieved using:
- **User-User Collaborative Filtering**
- **Item-Item Collaborative Filtering**
- Optionally, **Matrix Factorization** (e.g., using SVD)

---

## 🧠 Technologies Used

- Python 3
- Pandas & NumPy
- Scikit-learn
- Surprise (a Python scikit for recommender systems)
- Matplotlib/Seaborn (for data visualization)

---

## 🎞️ Dataset

Suggested Dataset:  
📂 [MovieLens 100K Dataset](https://grouplens.org/datasets/movielens/100k/)

Dataset includes:
- Users
- Movies
- Ratings (from 1 to 5)

---

## 📁 Project Structure